[#1][#2][#3] Use channel permission overwrites instead of roles
So far the biggest limitation to using role based permissions for channels was that every time a channel needed to be visible for members, an associated role needed to exist such that the role assignment would take care of channel visibility/writing permissions for the user. By making the command read and write to the channel's permission overwrites instead, the need for roles will be gone and hence the limitation of 250 roles per guild will be also vanish. The only apprehension here is that it is unknown (and as far as I could tell, undocumented) of the maximum of permission overwrites each channel can maintain. This would need to be tested, however, testing this could be tedious since the need to add visibility to channels for a large number of members has to be done manually (unless it is scripted to take the first 200-300 members in a guild and assigning them all to some test channel?).